Splitting up the Redux reducers into separate files
