+++ src/reducers/preparedQuestions.js
@@ -1,0 +1,10 @@
+import { SET_PREPARED_QUESTIONS } from '../actions';
+
+const preparedQuestions = (state = [], action) => {
+  if (action.type === SET_PREPARED_QUESTIONS) {
+    return action.payload;
+  }
+  return state;
+};
+
+export default preparedQuestions;
