+++ src/reducers/usedQuestions.js
@@ -1,0 +1,10 @@
+import { SET_USED_QUESTIONS } from '../actions';
+
+const usedQuestions = (state = [], action) => {
+  if (action.type === SET_USED_QUESTIONS) {
+    return action.payload;
+  }
+  return state;
+};
+
+export default usedQuestions;
